+++ planetarium/gui/gui_top.gd
@@ -17,22 +17,27 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # *****************************************************************************
-# See comments in ivoyager/gui_example/example_game_gui.gd to understand what's
-# going on here.
-
-extends Control
 class_name GUITop
+extends Control
 const SCENE := "res://planetarium/gui/gui_top.tscn"
+
+# Scenes instanced by IVProjectBuilder need SCENE constant above.
+#
+# An IVSelectionManager instance manages our current selection. To find this
+# instanace, various GUI widgets search up their ancestor tree for the first
+# node that has a "selection_manager" member.
 
 var selection_manager: IVSelectionManager
 
 onready var _SelectionManager_: Script = IVGlobal.script_classes._SelectionManager_
+
 
 func _project_init() -> void:
 	IVGlobal.connect("project_builder_finished", self, "_on_project_builder_finished")
 	IVGlobal.connect("system_tree_built_or_loaded", self, "_on_system_tree_built_or_loaded")
 	IVGlobal.connect("simulator_exited", self, "_on_simulator_exited")
 	hide()
+
 
 func _ready():
 	var style_box := StyleBoxFlat.new()
@@ -45,13 +50,16 @@ func _ready():
 	var set_date_time: Button = find_node("SetDateTime")
 	set_date_time.connect("pressed", $TimeSetPopup, "popup")
 
+
 func _on_project_builder_finished() -> void:
 	theme = IVGlobal.themes.main
+
 
 func _on_system_tree_built_or_loaded(is_new_game: bool) -> void:
 	if is_new_game:
 		selection_manager = _SelectionManager_.new()
 	show()
+
 
 func _on_simulator_exited() -> void:
 	hide()